fix: change logic decoratos
