+++ src/components/Calendar/Calendar.types.ts
@@ -1,14 +1,15 @@
-import { TypeTodo } from "components/TodoList/TodoItem/TodoItem.types";
+import { type TypeTodo } from "components/TodoList/TodoItem/TodoItem.types";
+
 import { type TypeYear } from "types/contextTypes";
 import { type TypeTargetDay } from "types/datePickerTypes";
 
 export interface PropsCalendar {
-  typeStart: "Mo" | "Su";
-  withHolidays: boolean;
+  typeStart?: "Mo" | "Su";
+  withHolidays?: boolean;
   targetDay?: TypeTargetDay | null;
   withRangePicker?: boolean;
   range?: TypeYear[] | null;
-  colorHoliday: "red" | "lightRed" | "apple";
+  colorHoliday?: "red" | "lightRed" | "apple";
   todos?: Record<string, TypeTodo[]>;
   setViewTodo?: (value: boolean) => void;
   handleClickRange?: (year: number, month: number, day: number) => void;
@@ -19,7 +20,7 @@ export interface PropsCalendar {
     day: number,
     type: string
   ) => void;
-  min: number;
-  max: number;
+  min?: number;
+  max?: number;
   ButtonClear?: React.ReactNode;
 }
