+++ src/components/DatePicker/DatePicker.types.ts
@@ -1,0 +1,7 @@
+export interface PropsDatePicker {
+  typeStart?: "Mo" | "Su";
+  withHolidays?: boolean;
+  colorHoliday?: "red" | "lightRed" | "apple";
+  min?: number;
+  max?: number;
+}
