+++ src/components/DateRangeCalendar/DateRangeCalendar.styled.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const StyledWrapperDateRange = styled.div`
+  position: relative;
+  width: 250px;
+  overflow: hidden;
+`;
+
+export { StyledWrapperDateRange };
