+++ src/components/DateRangeCalendar/DateRangeCalendar.types.ts
@@ -1,0 +1,7 @@
+export interface PropsDateRangeCalendar {
+  typeStart?: "Mo" | "Su";
+  withHolidays?: boolean;
+  colorHoliday?: "red" | "lightRed" | "apple";
+  min?: number;
+  max?: number;
+}
